Add Launcher class
